// src/utils/types/CvDetailsInterface.ts
export interface CvDetailsInterface {
  id: string;
  avatar: string;
  mail: string;
  firstName: string;
  lastName: string;
  education: string;
  courses: string;
  workExperience: string;
  portfolioUrls: string;
  bonusProjectUrls: string;
  phone: string;
  githubUsername: string;
  bio: string;
  courseCompletion: number;
  courseEngagment: number;
  projectDegree: number;
  teamProjectDegree: number;
  expectedTypeWork: string;
  targetWorkCity: string;
  expectedContractType: string;
  expectedSalary: number;
  canTakeApprenticeship: string;
  monthsOfCommercialExp: string;
  githubLink: string;
}